story_teller_v1/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile story_teller_v1/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-adk
    #   google-genai
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
authlib==1.6.0
    # via google-adk
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   google-adk
    #   uvicorn
cloudpickle==3.1.1
    # via google-cloud-aiplatform
cryptography==45.0.5
    # via authlib
docstring-parser==0.16
    # via google-cloud-aiplatform
fastapi==0.116.1
    # via google-adk
google-adk==1.7.0
    # via -r story_teller_v1/requirements.in
google-api-core[grpc]==2.25.1
    # via
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
google-api-python-client==2.176.0
    # via google-adk
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
    #   google-genai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-aiplatform[agent-engines]==1.104.0
    # via google-adk
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-bigquery==3.35.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.12.1
    # via google-cloud-aiplatform
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-secret-manager==2.24.0
    # via google-adk
google-cloud-speech==2.33.0
    # via google-adk
google-cloud-storage==2.19.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-cloud-trace==1.16.2
    # via
    #   google-cloud-aiplatform
    #   opentelemetry-exporter-gcp-trace
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.26.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
graphviz==0.21
    # via google-adk
greenlet==3.2.3
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
grpcio==1.73.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.73.1
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jsonschema==4.24.0
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
mcp==1.11.0
    # via google-adk
numpy==2.3.1
    # via shapely
opentelemetry-api==1.35.0
    # via
    #   google-adk
    #   google-cloud-logging
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-gcp-trace==1.9.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
opentelemetry-resourcedetector-gcp==1.9.0a0
    # via opentelemetry-exporter-gcp-trace
opentelemetry-sdk==1.35.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
opentelemetry-semantic-conventions==0.56b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   google-adk
    #   google-cloud-aiplatform
    #   google-genai
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   google-adk
    #   google-cloud-bigquery
python-dotenv==1.1.1
    # via
    #   google-adk
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via google-adk
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-adk
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   opentelemetry-resourcedetector-gcp
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shapely==2.1.1
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via google-adk
sse-starlette==2.4.1
    # via mcp
starlette==0.47.1
    # via
    #   fastapi
    #   google-adk
    #   mcp
tenacity==8.5.0
    # via
    #   google-adk
    #   google-genai
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   google-adk
    #   google-cloud-aiplatform
    #   google-genai
    #   opentelemetry-api
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzlocal==5.3.1
    # via google-adk
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via
    #   google-adk
    #   mcp
watchdog==6.0.0
    # via google-adk
websockets==15.0.1
    # via
    #   google-adk
    #   google-genai
zipp==3.23.0
    # via importlib-metadata
